refactor pages, templates and components
