+++ src/components/Bio.js
@@ -69,5 +69,3 @@ class Bio extends Component {
     );
   }
 }
-
-export default Bio;
