+++ src/components/Card.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import media from "../media-query";
+import { media } from "../ui";
 
 const CardWrapper = styled.div`
   display: flex;
