+++ src/components/index.js
@@ -1,0 +1,21 @@
+import { FullWidth } from "./FullWidth";
+import { Categories } from "./Categories";
+import { TagFilter } from "./TagFilter";
+import { Authors } from "./Bio";
+import DocListing from "./DocListing";
+import BlogListing from "./BlogListing";
+import { BlogTag, Tag } from "./Tags";
+import SearchEngineOptimization from "./SearchEngineOptimization";
+
+//import explicitly to list all shared components below.
+export {
+  Authors,
+  BlogTag,
+  BlogListing,
+  DocListing,
+  Categories,
+  FullWidth,
+  SearchEngineOptimization,
+  TagFilter,
+  Tag,
+};
