+++ src/pages/tags.js
@@ -3,7 +3,8 @@ import PropTypes from "prop-types";
 import kebabCase from "lodash/kebabCase";
 import { Helmet } from "react-helmet";
 import { Link, graphql } from "gatsby";
-import Layout from "../components/Layout";
+
+import { Layout } from "../ui";
 
 const TagPage = ({ pageContext, data, location }) => {
   const { group } = data.allMarkdownRemark;
