+++ src/templates/doc-main/Title.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const Title = styled.div`
+  color: #333333;
+  font-family: Equinor;
+  font-size: ${props => props.fontSize || 52}px;
+  letter-spacing: -0.1px;
+  line-height: 52px;
+  text-align: center;
+`;
