+++ src/ui/index.js
@@ -1,0 +1,6 @@
+import GlobalStyle from "./global-style";
+import style from "./style";
+import media from "./media-query";
+import Layout from "../components/layout/Layout";
+
+export { GlobalStyle, style, media, Layout };
